+++ post-formats/format.php
@@ -47,8 +47,17 @@
                 </section> <?php // end article section ?>
 
                 <aside class="article-supplement">
-                  <?php echo '<p id="figcaption_' . $id . '" class="wp-caption-text" itemprop="description">Headline image: ' . $img['caption'] .'</figcaption></figure>'; ?>
-                  <?php the_tags( '<p class="tag-titles"><span>' . __( 'Tags:', 'bonestheme' ) . '</span> ', ', ', '</p>' ); ?>
+                  <?php 
+                    echo '<p id="figcaption_' . $id . '" class="wp-caption-text" itemprop="description">Headline image: ' . $img['caption'] .'</figcaption></figure>';
+                    the_tags( '<p class="tag-titles"><span>' . __( 'Tags:', 'bonestheme' ) . '</span> ', ', ', '</p>' );
+                    /* Adds Jetpack "Like" button iframe back in.
+                    // @todo Uncomment if we need this crufty crap (don't forget to activate the JP likes module)
+                    if ( class_exists( 'Jetpack_Likes' ) ) {
+                      $gramm_likes = new Jetpack_Likes;
+                      echo $gramm_likes->post_likes( '' );
+                    } */
+                  ?>
+
                 </aside>
 
                 <footer class="article-footer">